fastapi_plan/template/{{cookiecutter.project_name}}/requirements.txt
aerich==0.5.3; python_version >= "3.7" and python_version < "4.0"
aiosqlite==0.16.1; python_version >= "3.7" and python_version < "4.0"
appdirs==1.4.4; python_full_version >= "3.6.2"
asyncpg==0.22.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
asynctest==0.13.0; python_version >= "3.5"
atomicwrites==1.4.0; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.4.0"
attrs==21.1.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.5"
bcrypt==3.2.0; python_version >= "3.6"
black==21.5b0; python_full_version >= "3.6.2"
certifi==2020.12.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cffi==1.14.5
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
click==7.1.2; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.5.0"
cryptography==3.4.7; python_version >= "3.6"
ddlparse==1.9.0; python_version >= "3.7" and python_version < "4.0"
dictdiffer==0.8.1; python_version >= "3.7" and python_version < "4.0"
dnspython==2.1.0; python_full_version >= "3.6.1" and python_version >= "3.6"
ecdsa==0.14.1; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
email-validator==1.1.2; python_full_version >= "3.6.1" and python_version >= "3.6"
fastapi==0.63.0; python_version >= "3.6"
flake8==3.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
h11==0.12.0; python_version >= "3.6"
idna==2.10; python_full_version >= "3.6.1" and python_version >= "3.6"
iso8601==0.1.14; python_version >= "3.7" and python_version < "4.0"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
more-itertools==8.7.0; python_version >= "3.5"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
packaging==20.9; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.5"
passlib==1.7.4
pathspec==0.8.1; python_full_version >= "3.6.2"
pluggy==0.13.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.5"
py==1.10.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.5"
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pydantic==1.8.1; python_full_version >= "3.6.1"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyparsing==2.4.7; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.4.0"
pypika-tortoise==0.1.0; python_version >= "3.7" and python_version < "4.0"
pytest==5.4.3; python_version >= "3.5"
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.5; python_version >= "3.7" and python_version < "4.0"
regex==2021.4.4; python_full_version >= "3.6.2"
requests==2.25.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
rsa==4.7.2; python_version >= "3.5" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.3.0"
starlette==0.13.6; python_version >= "3.6"
tenacity==7.0.0
toml==0.10.2; python_full_version >= "3.6.2"
tortoise-orm==0.17.2; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.0; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
uvicorn==0.13.4
wcwidth==0.2.5; python_version >= "3.5"
